@@ -1,6 +1,6 @@
 module bothoi
 
-go 1.17
+go 1.19
 
 require (
 	github.com/gorilla/websocket v1.5.0
@@ -10,11 +10,6 @@ require (
 )
 
 require (
-	github.com/bitly/go-simplejson v0.5.0 // indirect
-	github.com/dlclark/regexp2 v1.4.1-0.20201116162257-a2a8dda75c91 // indirect
-	github.com/dop251/goja v0.0.0-20211211112501-fb27c91c26ed // indirect
-	github.com/go-sourcemap/sourcemap v2.1.3+incompatible // indirect
 	github.com/jonas747/ogg v0.0.0-20161220051205-b4f6f4cf3757 // indirect
 	golang.org/x/sys v0.0.0-20211214234402-4825e8c3871d // indirect
-	golang.org/x/text v0.3.7 // indirect
 )
