@@ -11,14 +11,14 @@ type GatewayPayload struct {
 	Op gateway_opcode.GatewayOpcode `json:"op" mapstructure:"op"`
 	S  *uint64                      `json:"s,omitempty" mapstructure:"s"`
 	T  string                       `json:"t" mapstructure:"t"`
-	D  interface{}                  `json:"d" mapstructure:"d"`
+	D  any                          `json:"d" mapstructure:"d"`
 }
 
 type VoiceGatewayPayload struct {
 	Op voice_opcode.VoiceOpcode `json:"op" mapstructure:"op"`
 	S  *uint64                  `json:"s,omitempty" mapstructure:"s"`
 	T  string                   `json:"t" mapstructure:"t"`
-	D  interface{}              `json:"d" mapstructure:"d"`
+	D  any                      `json:"d" mapstructure:"d"`
 }
 
 func NewHeartbeat(s *uint64) GatewayPayload {
@@ -31,10 +31,10 @@ func NewHeartbeat(s *uint64) GatewayPayload {
 func NewIdentify() GatewayPayload {
 	return GatewayPayload{
 		Op: gateway_opcode.Identify,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"token":   config.BOT_TOKEN,
 			"intents": config.GATEWAY_INTENT,
-			"properties": map[string]interface{}{
+			"properties": map[string]any{
 				"$os":      "linux",
 				"$browser": "bothoi",
 				"$device":  "bothoi",
@@ -47,7 +47,7 @@ func NewIdentify() GatewayPayload {
 func NewResume(s *uint64, sessionId string) GatewayPayload {
 	return GatewayPayload{
 		Op: gateway_opcode.Resume,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"token":      config.BOT_TOKEN,
 			"session_id": sessionId,
 			"seq":        s,
@@ -58,7 +58,7 @@ func NewResume(s *uint64, sessionId string) GatewayPayload {
 func NewVoiceStateUpdate(guildId string, voiceId *string, mute, deaf bool) GatewayPayload {
 	return GatewayPayload{
 		Op: gateway_opcode.VoiceStateUpdate,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"guild_id":   guildId,
 			"channel_id": voiceId,
 			"self_mute":  mute,
@@ -70,7 +70,7 @@ func NewVoiceStateUpdate(guildId string, voiceId *string, mute, deaf bool) Gatew
 func NewVoiceIdentify(guildId, userId, sessionId, token string) VoiceGatewayPayload {
 	return VoiceGatewayPayload{
 		Op: voice_opcode.Identify,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"server_id":  guildId,
 			"user_id":    userId,
 			"session_id": sessionId,
@@ -90,9 +90,9 @@ func NewVoiceHeartbeat(s int64) VoiceGatewayPayload {
 func NewVoiceSelectProtocol(address string, port uint16, mode string) VoiceGatewayPayload {
 	return VoiceGatewayPayload{
 		Op: voice_opcode.SelectProtocol,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"protocol": "udp",
-			"data": map[string]interface{}{
+			"data": map[string]any{
 				"address": address,
 				"port":    port,
 				"mode":    mode,
@@ -104,7 +104,7 @@ func NewVoiceSelectProtocol(address string, port uint16, mode string) VoiceGatew
 func NewVoiceSpeaking(ssrc uint32) VoiceGatewayPayload {
 	return VoiceGatewayPayload{
 		Op: voice_opcode.Speaking,
-		D: map[string]interface{}{
+		D: map[string]any{
 			"speaking": 1 << 0,
 			"delay":    0,
 			"ssrc":     ssrc,
